GROK-19380: Diff Grok: Fixed docs
Fixed import for correct docs
